[Added] RequestCraftingSession, CancelCraftingSession [Added] CraftingTool, CraftingSession [Updated] dplay9
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@2113 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/draftschematic/DraftSchematic.h
@@ -71,7 +71,7 @@ public:
 
 	void setToolTab(unsigned int tab);
 
-	float getToolTab();
+	unsigned int getToolTab();
 
 	void setSize(unsigned int s);
 
@@ -172,7 +172,7 @@ public:
 
 	void setToolTab(unsigned int tab);
 
-	float getToolTab();
+	unsigned int getToolTab();
 
 	void setSize(unsigned int s);
 
@@ -249,7 +249,7 @@ public:
 
 	void setToolTab(unsigned int tab);
 
-	float getToolTab();
+	unsigned int getToolTab();
 
 	void setSize(unsigned int s);
